+++ jave-core/src/main/java/ws/schild/jave/DefaultFFMPEGLocator.java
@@ -24,8 +24,8 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * The default ffmpeg executable locator, which exports on disk the ffmpeg
@@ -37,13 +37,13 @@ import org.apache.commons.logging.LogFactory;
  */
 public class DefaultFFMPEGLocator extends FFMPEGLocator {
 
-    private final static Log LOG = LogFactory.getLog(FFMPEGExecutor.class);
+    private static final Logger LOG = LoggerFactory.getLogger(DefaultFFMPEGLocator.class);
 
     /**
      * Trace the version of the bundled ffmpeg executable. It's a counter: every
      * time the bundled ffmpeg change it is incremented by 1.
      */
-    private static final String MY_EXE_VERSION = "2.5.1-SNAPSHOT";
+    private static final String MY_EXE_VERSION = "2.6.0-SNAPSHOT";
 
     /**
      * The ffmpeg executable file path.
@@ -58,18 +58,18 @@ public class DefaultFFMPEGLocator extends FFMPEGLocator {
         String os = System.getProperty("os.name").toLowerCase();
         boolean isWindows = os.contains("windows");
         boolean isMac = os.contains("mac");
-        LOG.debug("Os name is <"+os+"> isWindows: "+isWindows+" isMac: "+isMac);
+        LOG.debug("Os name is <{}> isWindows: {} isMac: {}", os, isWindows, isMac);
 
         // Dir Folder
         File dirFolder = new File(System.getProperty("java.io.tmpdir"), "jave/");
         if (!dirFolder.exists())
         {
-            LOG.debug("Creating jave temp folder to place executables in <"+dirFolder.getAbsolutePath()+">");
+            LOG.debug("Creating jave temp folder to place executables in <{}>", dirFolder.getAbsolutePath());
             dirFolder.mkdirs();
         }
         else
         {
-            LOG.debug("Jave temp folder exists in <"+dirFolder.getAbsolutePath()+">");
+            LOG.debug("Jave temp folder exists in <{}>", dirFolder.getAbsolutePath());
         }
 
         // -----------------ffmpeg executable export on disk.-----------------------------
@@ -78,17 +78,17 @@ public class DefaultFFMPEGLocator extends FFMPEGLocator {
 
         //File
         File ffmpegFile = new File(dirFolder, "ffmpeg-" + arch +"-"+MY_EXE_VERSION+ suffix);
-        LOG.debug("Executable path: "+ffmpegFile.getAbsolutePath());
+        LOG.debug("Executable path: {}", ffmpegFile.getAbsolutePath());
 
         //Check the version of existing .exe file
         if (ffmpegFile.exists())
         {
             // OK, already present
-            LOG.debug("Executable exists in <"+ffmpegFile.getAbsolutePath()+">");
+            LOG.debug("Executable exists in <{}>", ffmpegFile.getAbsolutePath());
         }
         else
         {
-            LOG.debug("Need to copy executable to <"+ffmpegFile.getAbsolutePath()+">");
+            LOG.debug("Need to copy executable to <{}>", ffmpegFile.getAbsolutePath());
             copyFile("ffmpeg-" + arch + suffix, ffmpegFile);
         }
 
@@ -103,13 +103,13 @@ public class DefaultFFMPEGLocator extends FFMPEGLocator {
                 });
             } catch (IOException e)
             {
-                LOG.error(e);
+                LOG.error("Error setting executable via chmod", e);
             }
         }
 
         // Everything seems okay
         path = ffmpegFile.getAbsolutePath();
-        LOG.debug("ffmpeg executable found: "+path);
+        LOG.debug("ffmpeg executable found: {}", path);
     }
 
     @Override
@@ -125,37 +125,44 @@ public class DefaultFFMPEGLocator extends FFMPEGLocator {
      * @throws RuntimeException If an unexpected error occurs.
      */
     private void copyFile(String path, File dest) {
-        String resourceName= "nativebin/" + path;
+       String resourceName= "nativebin/" + path;
         try
         {
-            LOG.debug("Copy from resource <"+resourceName+"> to target <"+dest.getAbsolutePath()+">");
+            LOG.debug("Copy from resource <{}> to target <{}>", resourceName, dest.getAbsolutePath());
             InputStream is= getClass().getResourceAsStream(resourceName);
+            if (is == null)
+            {
+                // Use this for Java 9+ only if required
+                resourceName= "ws/schild/jave/nativebin/" + path;
+                LOG.debug("Alternative copy from SystemResourceAsStream <{}> to target <{}>", resourceName, dest.getAbsolutePath());
+                is= ClassLoader.getSystemResourceAsStream(resourceName);
+            }
             if (is != null)
             {
                 if (copy(is, dest.getAbsolutePath()))
                 {
                     if (dest.exists())
                     {
-                        LOG.debug("Target <"+dest.getAbsolutePath()+"> exists");
+                        LOG.debug("Target <{}> exists", dest.getAbsolutePath());
                     }
                     else
                     {
-                        LOG.fatal("Target <"+dest.getAbsolutePath()+"> does not exist");
+                        LOG.error("Target <{}> does not exist", dest.getAbsolutePath());
                     }
                 }
                 else
                 {
-                    LOG.fatal("Copy resource to target <"+dest.getAbsolutePath()+"> failed");
+                    LOG.error("Copy resource to target <{}> failed", dest.getAbsolutePath());
                 }
             }
             else
             {
-                LOG.fatal("Could not find ffmpeg platform executable in resources for <"+resourceName+">");
+                LOG.error("Could not find ffmpeg platform executable in resources for <{}>", resourceName);
             }
         }
         catch (NullPointerException ex)
         {
-            LOG.error("Could not find ffmpeg executable for "+resourceName+" is the correct platform jar included?");
+            LOG.error("Could not find ffmpeg executable for {} is the correct platform jar included?", resourceName);
             throw ex;
         }
     }
@@ -175,7 +182,7 @@ public class DefaultFFMPEGLocator extends FFMPEGLocator {
             Files.copy(source, Paths.get(destination), StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException ex)
         {
-            LOG.fatal("Cannot write file " + destination, ex);
+            LOG.error("Cannot write file " + destination, ex);
             success = false;
         }